use uint32_t instead of size_t for Def::gid_
* makes hashing slightly faster and constistent on 32 bit archs

+++ thorin/def.cpp
@@ -57,7 +57,7 @@ bool check_same_sorted_ops(Def::Sort sort, Defs ops) {
  * misc
  */
 
-size_t Def::gid_counter_ = 1;
+uint32_t Def::gid_counter_ = 1;
 
 Def::Sort Def::sort() const {
     if (auto t = type()) {
